+++ src/lib/sound/index.ts
@@ -1,0 +1,8 @@
+
+const sound = require("sound-play");
+
+
+const fs = require("fs");
+const os = require("os");
+const spawn = require("child_process").spawn;
+
