+++ src/lib/test/index.ts
@@ -1,0 +1,11 @@
+interface ITrace {
+  (label: any):any;
+}
+
+export const trace: ITrace = (label) => (value) => {
+  console.log(
+    `${label}: ${{...value}}`
+  );
+  return {...value};
+};
+
